# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "https://localhost:3000"
info:
  title: WASAPhoto
  description: Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
security:
  - BearerAuth: []

paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      security: []
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
            description: User log-in action successful
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    identifier:
                      type: string
                      example: "abcdef012345"
                      minLength: 3
                      maxLength: 16
  /users/name:
    put:
      summary: User can update their name
      operationId: setMyUserName
      
      tags: ["user"]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: "#/components/schemas/plainString"
      responses:
        '200':
          description: Name set successfully
          content:
              application/json:
                schema:
                  type: object
                  properties:
                    updated_name:
                      $ref: '#/components/schemas/plainString'
        '401':
          description: Unauthorized - User not authenticated
  /photos/:
    post:
      summary: Upload Photo
      operationId: uploadPhoto
      tags: ["photo"]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              additionalProperties: false
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  minLength: 0
                  maxLength: 10000000
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Bad Request - Invalid file format or size
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        '401':
          description: Unauthorized - User not authenticated    
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString' 
    delete:
      summary: Delete Photo
      operationId: deletePhoto
      tags: ["photo"]
      parameters:
        - in: query
          name: photoId
          schema:
            $ref: "#/components/schemas/idString"
          required: true
      security:
        - BearerAuth: []  
      responses:
        '204':
          description: Photo deleted successfully
        '400':
          description: Bad Request - Invalid photo ID or photo not found

  /users/{userId}/follow/:
    parameters:
            - name: userId
              in: path
              required: true
              schema:
                $ref: "#/components/schemas/idString"
    post:
      summary: Logged in user follows another
      operationId: followUser
      tags: ["user"]
      
      security:
        - BearerAuth: []
      responses:
        '204':
          description: User followed successfully
        '400':
          description: Bad Request - Invalid user ID or user not found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        '401':
          description: Unauthorized - User not authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        '403':
          description: User does not have the rights
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/plainString'
          
    delete:
      summary: Logged in user unfollows another
      operationId: unfollowUser
      tags: ["user"]
      security:
        - BearerAuth: []
      responses:
        '204':
          description: User unfollowed successfully
        '400':
          description: Bad Request - Invalid user ID or user not found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        '401':
          description: Unauthorized - User not authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        '403':
          description: User does not have the rights
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/plainString'
        
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/plainString'
  /users/{userId}/ban/:
    parameters:
          - name: userId
            in: path
            required: true
            schema:
              $ref: "#/components/schemas/idString"
    post:
      summary: Logged in user bans another user
      operationId: banUser
      tags: ["user"]
      security:
        - BearerAuth: []
      
      
      responses:
        '204':
          description: User banned successfully
        '400':
          description: Bad Request - Invalid user ID or user not found
        '401':
          description: Unauthorized - User not authenticated
    delete:
        summary: Logged in user unbans another user
        operationId: unbanUser
        tags: ["user"]
        security:
        - BearerAuth: []
        responses:
          '204':
            description: User unbanned successfully
          '400':
            description: Bad Request - Invalid user ID or user not found
          '401':
            description: Unauthorized - User not authenticated
  /users/{userId}:
      get:
        summary: Get User Profile
        operationId: getUserProfile
        tags: ["user"]
        security: []
        parameters:
          - name: userId
            in: path
            required: true
            schema:
              $ref: "#/components/schemas/idString"
        responses:
          '200':
            description: User profile retrieved successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    userId:
                      type: string
                      minLength: 0
                      maxLength: 16
                    username:
                      type: string
                      minLength: 3
                      maxLength: 16
                    photos:
                      type: array
                      minItems: 0
                      maxItems: 50
                      items:
                        $ref: '#/components/schemas/Photo'
                    followers:
                      type: array
                      minItems: 0
                      maxItems: 50
                      items:
                        type: object
                        properties:
                          userId:
                            type: string
                            minLength: 0
                            maxLength: 16
                          username:
                            type: string
                            minLength: 3
                            maxLength: 16
                    following:
                      type: array
                      minItems: 0
                      maxItems: 50
                      items:
                        type: object
                        properties:
                          userId:
                            type: string
                            minLength: 0
                            maxLength: 16
                          username:
                            type: string
                            minLength: 3
                            maxLength: 16
          '404':
            description: User not found
  /stream/:
    get:
      summary: User gets stream
      operationId: getMyStream
      tags: ["photo"]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Stream retrieved successfully
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 50

                items:
                  $ref: '#/components/schemas/Photo'
  
  /photos/{photoId}/like/:
    parameters:
            - name: photoId
              in: path
              required: true
              schema:
                $ref: "#/components/schemas/idString"
    post:
      summary: Like Photo
      operationId: likePhoto
      tags: ["photo"]
      
      security:
        - BearerAuth: []  
      responses:
        '204':
          description: Photo liked successfully
        '400':
          description: Bad Request - Invalid photo ID or photo not found
        '401':
          description: Unauthorized - User not authenticated
    delete:
      summary: Unlike Photo
      operationId: unlikePhoto
      tags: ["photo"]
      security:
        - BearerAuth: []  
      responses:
        '204':
          description: Photo unliked successfully
        '400':
          description: Bad Request - Invalid photo ID or photo not found
        '401':
          description: Unauthorized - User not authenticated

  /comments:
    post:
      summary: Comment on Photo
      operationId: commentPhoto
      tags: ["comment"]
      parameters:
        - in: query
          name: photoId
          schema:
            $ref: "#/components/schemas/idString"
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                comment:
                  $ref: "#/components/schemas/plainString"
      security:
        - BearerAuth: []  
      responses:
        '200':
          description: Comment added successfully
          content:
              application/json:
                schema:
                  type: object
                  additionalProperties: false
                  properties:
                    userId:
                      $ref: "#/components/schemas/plainString"
                    id:
                      $ref: "#/components/schemas/plainString"
                    imageId:
                      $ref: "#/components/schemas/plainString"
                    text:
                      $ref: "#/components/schemas/plainString"
                    CreateTime:
                      $ref: "#/components/schemas/plainString"
        '400':
          description: Bad Request - Invalid photo ID or photo not found
        '401':
          description: Unauthorized - User not authenticated
    delete:
      summary: Remove Comment from Photo
      operationId: uncommentPhoto
      tags: ["photo"]
      parameters:
        - in: query
          name: commentId
          schema:
            $ref: "#/components/schemas/idString"
          required: true
      security:
        - BearerAuth: []  
      responses:
        '204':
          description: Comment removed successfully
        '400':
          description: Bad Request - Invalid photo ID, photo not found, or comment not found


components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string
      description: The user identifier (User id) obtained after logging in. Include it in the Authorization header.
  schemas:
    plainString:
      type: string
      pattern: '^[a-zA-Z0-9]$' 
      maxLength: 100 
    idString:
      type: string
      pattern: '^[0-9]+$'
      maxLength: 30
    Photo:
      type: object
      additionalProperties: false
      properties:
        image:
          type: string
          pattern: '^[a-zA-Z0-9]$' 
          maxLength: 100 
        photoId:
          type: string
          pattern: '^[a-zA-Z0-9]$' 
          maxLength: 100 
        uploadDate:
          type: string
          pattern: '^[a-zA-Z0-9]$' 
          maxLength: 100 
        likes:
          type: integer
        comments:
          type: array
          minItems: 0
          maxItems: 50
          items:
            type: string
            pattern: '^[a-zA-Z0-9]$' 
            maxLength: 400 